==> eda/01_data_cleaning.Rnw <==
\documentclass{article}

\begin{document}
\SweaveOpts{concordance=TRUE}

<<load library>>=
library(haven)
library(tidyverse)
library(miscset)
library(xtable)
@

<<load-data>>=
# Female data
female_data <- read_por("../data/03212-0001-Data.por")
female_variable <- colnames(female_data)

length(grep(pattern = "^90-", female_data$CODENUM) )
length(grep(pattern = "^91-", female_data$CODENUM) )
@

<<data-cleanup>>=
# add cohort
female_data[grep(pattern = "^90-", female_data$CODENUM), "COHORT"] <- 90
female_data[grep(pattern = "^91-", female_data$CODENUM), "COHORT"] <- 91

# relation status
relation2_90 <- which((female_data$COHORT == 90) & (female_data$MARSTAT2 != 1))
relation2_91 <- which((female_data$COHORT == 91) & (female_data$RELSTAT2 != 1))
no_relation2_90 <- which((female_data$COHORT == 90) & (female_data$MARSTAT2 == 1))
no_relation2_91 <- which((female_data$COHORT == 91) & (female_data$RELSTAT2 == 1))
RELATION2 <- rep(NA, nrow(female_data))
RELATION2[c(relation2_90, relation2_91)] <- 1
RELATION2[c(no_relation2_90, no_relation2_91)] <- 0
female_data <- cbind(female_data, RELATION2)

# sexual experience from age 14 to the present
  # intercourse occurred 
no_forced_sex <- female_data$SEXACTS == 1 & female_data$FORCESI == 1 & 
  female_data$PRESSSP == 1 & female_data$FORCESP == 1 & female_data$AUTHSP == 1 &
  female_data$PRESSSI == 1 & female_data$AUTHSI == 1 & female_data$DRUGSI == 1
forced_sex <- female_data$SEXACTS == 2 | female_data$FORCESI == 2 | 
  female_data$PRESSSP == 2 | female_data$FORCESP == 2 | female_data$AUTHSP == 2 |
  female_data$PRESSSI == 2 | female_data$AUTHSI == 2 | female_data$DRUGSI == 2
FSEX_AFT14 <- rep(NA, nrow(female_data))
FSEX_AFT14[no_forced_sex] <- 0
FSEX_AFT14[forced_sex] <- 1
female_data <- cbind(female_data, FSEX_AFT14)
  # intercourse didnt occur
no_forced_sex_attempt <- female_data$ATTEMPT == 1 & female_data$DRUGATT == 1
forced_sex_attempt <- female_data$ATTEMPT == 2 | female_data$DRUGATT == 2
FSEXATT_AFT14 <- rep(NA, nrow(female_data))
FSEXATT_AFT14[no_forced_sex_attempt] <- 0
FSEXATT_AFT14[forced_sex_attempt] <- 1
female_data <- cbind(female_data, FSEXATT_AFT14)

# sexual experience prior to age 14
sex_bef14 <- female_data$SHOORGAN != 1 | female_data$FONDLED != 1 | 
  female_data$ATTEMPSI != 1 | female_data$SI != 1
no_sex_bef14 <- female_data$SHOORGAN == 1 & female_data$FONDLED == 1 &
  female_data$ATTEMPSI == 1 & female_data$SI == 1
SEX_BEF14 <- rep(NA, nrow(female_data))
SEX_BEF14[no_sex_bef14] <- 0
SEX_BEF14[sex_bef14] <- 1
female_data <- cbind(female_data, SEX_BEF14)

# drug use including marijuana or other drugs of the respondent
DRUG1 <- female_data$MJFREQ != 1 | female_data$OTHERRX != 1 # 1 for drug use, 0 for none
DRUG2 <- female_data$MJFREQ2 != 1 | female_data$OTHERRX2 != 1
DRUG3 <- female_data$MJFREQ3 != 1 | female_data$OTHERRX3 != 1
DRUG4 <- female_data$MJFREQ4 != 1 | female_data$OTHERRX4 != 1
DRUG5 <- female_data$MJFREQ5 != 1 | female_data$OTHERRX5 != 1
female_data <- cbind(female_data, DRUG1, DRUG2, DRUG3, DRUG4, DRUG5)

# # dating frequency during the past year 
# dating90 <- which((female_data$DATING2 != 1) & (female_data$COHORT == 90)) # 1 if dating at least once and in 90 cohort
# no_dating90 <- which((female_data$DATING2 == 1) & (female_data$COHORT == 90)) 
# dating91 <- which((female_data$DATEBEH2 == 3) | (female_data$DATEBEH2 == 4) & 
#                     (female_data$COHORT == 91)) 
# no_dating91 <- which((female_data$DATEBEH2 == 1) | (female_data$DATEBEH2 == 2) & (female_data$COHORT == 91)) 
# DATING2_comb <- rep(NA, nrow(female_data))
# DATING2_comb[c(dating90, dating91)] <- 1
# DATING2_comb[c(no_dating90, no_dating91)] <- 0
# female_data <- cbind(female_data, DATING2_comb)

@

<<impute baseline covariates>>=
# impute based on the raw data, otherwise remove any individuals with missing 
# value in the first survey

# sexual orientation
NA_sxorient <- which(is.na(female_data$SXORIENT))
apply(female_data[NA_sxorient, c("SXORIEN2", "SXORIEN3", "SXORIEN4", "SXORIEN5")], 2, 
      function(t){sum(is.na(t))})
female_data[NA_sxorient, "SXORIENT"] <- female_data[NA_sxorient, "SXORIEN2"]
@

<<final-table>>=
female_final <- female_data %>%
  # cohort 
  mutate(cohort = COHORT) %>%
  # add treatment intervention A
  mutate(victim1 = (COOCCURA != 1)) %>%
  mutate(victim2 = (COOCCUR2 != 1)) %>%
  mutate(victim3 = (COOCCUR3 != 1)) %>%
  mutate(victim4 = (COOCCUR4 != 1)) %>%
  # add suicidal L10:
  mutate(suicide1 = (SUICIDE != 1)) %>% # 1 if suicidal, 0 if not
  mutate(suicide2 = (SUICIDE2 != 1)) %>%
  mutate(suicide3 = (SUICIDE3 != 1)) %>%
  mutate(suicide4 = (SUICIDE4 != 1)) %>% 
  mutate(suicide5 = (SUICIDE5 != 1)) %>%
  # add outcome Y
  mutate(Y = as.numeric(suicide2) + as.numeric(suicide3) +
           as.numeric(suicide4) + as.numeric(suicide5)) %>%
  # race W_1: 
  mutate(race = RACE) %>%
  # sexual orientation W_2: 
  mutate(sxorient = (SXORIENT != 1)) %>% # 1 if bisexual/homosexual/not sure, 0 if heterosexual
  # consent sex (age 14 - present) W:_3 
  mutate(consent = CONSENT) %>%
  # forced sex (age 14 - present) W_4:
  mutate(forced = FSEX_AFT14) %>%
  mutate(forced_att = FSEXATT_AFT14) %>%
  # sexual experience before 14 W_5
  mutate(sex_bef14 = SEX_BEF14) %>%
  # relation status L_1
  mutate(relation1 = (MARSTAT != 1)) %>% # 1 for in-relation or separated, 0 for single
  mutate(relation2 = RELATION2) %>%
  mutate(relation3 = (RELSTAT3 != 1)) %>%
  mutate(relation4 = (RELSTAT4 != 1)) %>%
  mutate(relation5 = (RELSTAT5 != 1)) %>%
  # religious preference L_2
  mutate(religion1 = (RELINFL != 1)) %>% # 1 if religion has at least some 
  #influence on the way the subject choose to spend their time each day
  mutate(religion2 = (RELINFL2 != 1)) %>%
  mutate(religion3 = (RELINFL3 != 1)) %>%
  mutate(religion4 = (RELINFL4 != 1)) %>%
  mutate(religion5 = (RELINFL5 != 1)) %>%
  # negative psychological quality L_3
  mutate(psystress1 = PSYSTRES) %>% # continuous
  mutate(psystress2 = PSYSTRS2) %>%
  mutate(psystress3 = PSYSTRS3) %>%
  mutate(psystress4 = PSYSTRS4) %>%
  mutate(psystress5 = PSYSTRS5) %>%
  # positive psychological quality L_4
  mutate(psywell1 = PSYWELL) %>% # continuous
  mutate(psywell2 = PSYWELL2) %>%
  mutate(psywell3 = PSYWELL3) %>%
  mutate(psywell4 = PSYWELL4) %>%
  mutate(psywell5 = PSYWELL5) %>%
  # number of women the respondent know who had been sexually victimized L_5
  mutate(victfrnd1 = (VICTFRND != 1)) %>% # 1 for at least one, 0 for none
  mutate(victfrnd2 = (VCTFRND2 != 1)) %>%
  mutate(victfrnd3 = (VCTFRND3 != 1)) %>%
  mutate(victfrnd4 = (VCTFRND4 != 1)) %>%
  mutate(victfrnd5 = (NUMVICT5 != 1)) %>%
  # drug use of the respondent L_6
  mutate(drug1 = DRUG1) %>% # 1 for drug use, 0 for none
  mutate(drug2 = DRUG2) %>%
  mutate(drug3 = DRUG3) %>%
  mutate(drug4 = DRUG4) %>%
  mutate(drug5 = DRUG5) %>%
  # alcohol use of the respondent L_7
  mutate(alcohol1 = (DRNKFREQ != 1)) %>% # 1 for alcohol use, 0 for none
  mutate(alcohol2 = (DRNKFRQ2 != 1)) %>%
  mutate(alcohol3 = (DRNKFRQ3 != 1)) %>%
  mutate(alcohol4 = (DRNKFRQ4 != 1)) %>%
  mutate(alcohol5 = (DRNKFRQ5 != 1))
  # # dating frequency during the past year L_8
  # # no info at yr 1 since we dont care about the year before yr 1
  # mutate(dating2 = DATING2_comb) %>% # 1 if occa/freq date, 0 if (almost) never 
  # mutate(dating3 = (DATEBEH3 == 3 | DATEBEH3 == 4)) %>%
  # mutate(dating4 = (DATEBEH4 == 3 | DATEBEH4 == 4)) %>%
  # mutate(dating5 = (DATEBEH5 == 3 | DATEBEH5 == 4))

# export CSV file
SCM_nodes <- c("victim1", "victim2", "victim3", "victim4", "suicide1", 
               "suicide2", "suicide3", "suicide4", "suicide5", "Y", "race", 
               "sxorient", "consent", "forced", "forced_att", "sex_bef14", 
               "relation1", "relation2", "relation3", "relation4", "relation5", 
               "religion1", "religion2", "religion3", "religion4", "religion5", 
               #"psystress1", "psystress2", "psystress3", "psystress4", 
               #"psystress5", 
               "psywell1", "psywell2", "psywell3", "psywell4", 
               "psywell5", "victfrnd1", "victfrnd2", "victfrnd3", "victfrnd4", 
               "victfrnd5", "drug1", "drug2", "drug3", "drug4", "drug5", 
               "alcohol1", "alcohol2", "alcohol3", "alcohol4", "alcohol5")
female_SCM <- female_final[, c("CODENUM", "cohort", SCM_nodes)]
write.table(female_SCM, "../data/female_SCM.csv")
save(female_SCM, file = "../data/female_SCM.RData")
@

<<load-data>>=
load("../data/female_SCM.RData")
@

\end{document}
