==> eda/02_EDA.Rnw <==
\documentclass{article}

\begin{document}
\SweaveOpts{concordance=TRUE}

<<load>>=
# load library
library(haven)
library(tidyverse)
library(grid)
library(gridExtra)

# load data
female_data <- read_por("../data/03212-0001-Data.por")
load("../data/female_SCM.RData")

source("../src/utility.R")
@

<<>>=
female_new <- female_data %>%
  mutate(victim1_bin = (COOCCURA != 1)) %>%
  mutate(victim2_bin = (COOCCUR2 != 1)) %>%
  mutate(victim3_bin = (COOCCUR3 != 1)) %>%
  mutate(victim4_bin = (COOCCUR4 != 1)) %>%
  mutate(SUICIDE1 = (SUICIDE != 1)) %>%
  mutate(SUICIDE2 = (SUICIDE2 != 1)) %>%
  mutate(SUICIDE3 = (SUICIDE3 != 1)) %>%
  mutate(SUICIDE4 = (SUICIDE4 != 1)) %>% # 1 if suicidal, 0 if not
  mutate(SUICIDE5 = (SUICIDE5 != 1)) %>%
  # 1 if engaged or married, 0 if single or divorced
  mutate(MARSTAT_bin = (MARSTAT == 3 | MARSTAT == 4)) %>%
  # 1 if has religious belief, o.w. 0
  mutate(RELIG_bin = (RELIG != 5)) %>%
  # 1 if depressed, o.w. 0
  mutate(DEPRESS1_bin = (DEPRESS != 1)) %>%
  mutate(DEPRESS2_bin = (DEPRESS2 != 1)) %>%
  mutate(DEPRESS3_bin = (DEPRESS3 != 1)) %>%
  mutate(DEPRESS4_bin = (DEPRESS4 != 1)) %>%
  mutate(DEPRESS5_bin = (DEPRESS5 != 1))
@

<<EDA-suicide, fig = T>>=
p_suicide1 <- ggplot(female_new, aes(x = SUICIDE1, fill = victim1_bin)) + 
  geom_bar() + xlab("Year 1")  + scale_nice + theme_bottom

suicide_legend <- Glegend(p_suicide1)

p_suicide2 <- ggplot(female_new, aes(x = SUICIDE2, fill = victim2_bin)) + 
  geom_bar() + xlab("Year 2") + scale_nice + theme_no_legend

p_suicide3 <- ggplot(female_new, aes(x = SUICIDE3, fill = victim3_bin)) + 
  geom_bar() + xlab("Year 3") + scale_nice + theme_no_legend

p_suicide4 <- ggplot(female_new, aes(x = SUICIDE4, fill = victim4_bin)) + 
  geom_bar() + xlab("Year 4") + scale_nice + theme_no_legend

p_outcome <- grid.arrange(arrangeGrob(p_suicide1 + theme(legend.position = "none"), 
                          p_suicide2, p_suicide3, p_suicide4, nrow = 2, ncol = 2), 
                            suicide_legend, nrow = 2, heights = c(10, 1),
                            top = textGrob("Indicator of suicide-related event per year", 
                                           gp = gpar(fontsize = 18, font = 3)))
ggsave(file = "../figure/dist-outcome.png", p_outcome)
@

<<eda-suicide，echo = F, eval = F>>=
female_new %>%
  filter(SUICIDE1 == 1) %>%
  filter(SUICIDE2 == 1) %>%
  filter(SUICIDE3 == 1) %>%
  filter(SUICIDE4 == 1) %>%
  summarise(n())
# 8
female_new %>%
  filter(SUICIDE1 == 0) %>%
  filter(SUICIDE2 == 0) %>%
  filter(SUICIDE3 == 0) %>%
  filter(SUICIDE4 == 0) %>%
  summarise(n())
# 754
@

<<eda-confounders, fig = T>>=
p_depress1 <- ggplot(female_new, aes(x = DEPRESS1_bin, fill = victim1_bin)) + 
  geom_bar() + xlab("Year 1")  + scale_nice + theme_bottom

depress_legend <- Glegend(p_depress1)

p_depress2 <- ggplot(female_new, aes(x = DEPRESS2_bin, fill = victim2_bin)) + 
  geom_bar() + xlab("Year 2") + scale_nice + theme_no_legend

p_depress3 <- ggplot(female_new, aes(x = DEPRESS3_bin, fill = victim3_bin)) + 
  geom_bar() + xlab("Year 3") + scale_nice + theme_no_legend

p_depress4 <- ggplot(female_new, aes(x = DEPRESS4_bin, fill = victim4_bin)) + 
  geom_bar() + xlab("Year 4") + scale_nice + theme_no_legend

p_depress <- grid.arrange(arrangeGrob(p_depress1 + 
                                           theme(legend.position = "none"), 
                p_depress2, p_depress3, p_depress4, nrow = 2, ncol = 2), 
                            depress_legend, nrow = 2, heights = c(10, 1),
                            top = textGrob("Depression Level Classified by Victimization", 
                                           gp = gpar(fontsize = 18, font = 3)))
ggsave(file = "../figure/dist-depression.png", p_depress)

p_confounder1 <- ggplot(female_new, aes(x = MARSTAT_bin, fill = victim1_bin)) + 
  geom_bar() + xlab("Relationship Status")  + scale_relation + theme_bottom

confounder_legend <- Glegend(p_confounder1)

p_confounder2 <- ggplot(female_new, aes(x = RELIG_bin, fill = victim2_bin)) + 
  geom_bar() + xlab("Religion Preference") + scale_religion + theme_no_legend

p_confounder <- grid.arrange(arrangeGrob(p_confounder1 + 
                                           theme(legend.position = "none"), 
                p_confounder2, nrow = 1, ncol = 2), 
                            confounder_legend, nrow = 2, heights = c(4, 1),
                            top = textGrob("Confounders Classified by Victimization", 
                                           gp = gpar(fontsize = 18, font = 3)))
ggsave(file = "../figure/dist-confounder.png", p_confounder)
@


i.	Analysis sample
1.	# of observations (individuals) meeting inclusion criteria, reasons for exclusion
2.	Consider a flow diagram
<<>>=
psywell1_high <- quantile(as.numeric(female_SCM$psywell1), 0.975, na.rm = T)
psywell2_high <- quantile(as.numeric(female_SCM$psywell2), 0.975, na.rm = T)
psywell3_high <- quantile(as.numeric(female_SCM$psywell3), 0.975, na.rm = T)
psywell4_high <- quantile(as.numeric(female_SCM$psywell4), 0.975, na.rm = T)
psywell5_high <- quantile(as.numeric(female_SCM$psywell5), 0.975, na.rm = T)
psywell1_low <- quantile(as.numeric(female_SCM$psywell1), 0.025, na.rm = T)
psywell2_low <- quantile(as.numeric(female_SCM$psywell2), 0.025, na.rm = T)
psywell3_low <- quantile(as.numeric(female_SCM$psywell3), 0.025, na.rm = T)
psywell4_low <- quantile(as.numeric(female_SCM$psywell4), 0.025, na.rm = T)
psywell5_low <- quantile(as.numeric(female_SCM$psywell5), 0.025, na.rm = T)
psywell_outlier1 <- which(female_SCM$psywell1 > psywell1_high & female_SCM$psywell2 > psywell2_high & 
                          female_SCM$psywell3 > psywell3_high & female_SCM$psywell4 > psywell3_high &
                          female_SCM$psywell5 > psywell5_high)
psywell_outlier2 <- which(female_SCM$psywell1 < psywell1_low & female_SCM$psywell2 < psywell2_low & 
                          female_SCM$psywell3 < psywell3_low & female_SCM$psywell4 < psywell3_low &
                          female_SCM$psywell5 < psywell5_low)

psystress1_high <- quantile(as.numeric(female_SCM$psystress1), 0.975, na.rm = T)
psystress2_high <- quantile(as.numeric(female_SCM$psystress2), 0.975, na.rm = T)
psystress3_high <- quantile(as.numeric(female_SCM$psystress3), 0.975, na.rm = T)
psystress4_high <- quantile(as.numeric(female_SCM$psystress4), 0.975, na.rm = T)
psystress5_high <- quantile(as.numeric(female_SCM$psystress5), 0.975, na.rm = T)
psystress1_low <- quantile(as.numeric(female_SCM$psystress1), 0.025, na.rm = T)
psystress2_low <- quantile(as.numeric(female_SCM$psystress2), 0.025, na.rm = T)
psystress3_low <- quantile(as.numeric(female_SCM$psystress3), 0.025, na.rm = T)
psystress4_low <- quantile(as.numeric(female_SCM$psystress4), 0.025, na.rm = T)
psystress5_low <- quantile(as.numeric(female_SCM$psystress5), 0.025, na.rm = T)
psystress_outlier1 <- which(female_SCM$psystress1 > psystress1_high & female_SCM$psystress2 > psystress2_high & 
                female_SCM$psystress3 > psystress3_high & female_SCM$psystress4 > psystress3_high &
                female_SCM$psystress5 > psystress5_high)
psystress_outlier2 <- which(female_SCM$psystress1 <psystress1_low & female_SCM$psystress2 < psystress2_low & 
                female_SCM$psystress3 < psystress3_low & female_SCM$psystress4 < psystress3_low &
                female_SCM$psystress5 < psystress5_low)

outliers <- c(psywell_outlier1, psywell_outlier2, psystress_outlier1, psystress_outlier2)
female_outliers <- female_SCM[-outliers, ]
@
ii.	Descriptive
1.	number of individuals following each regime of interest

<<>>=
vic_1111 <- sum((female_SCM$victim1 == 1) & (female_SCM$victim2 == 1) & 
                  (female_SCM$victim3 == 1) & (female_SCM$victim4 == 1), 
                na.rm = T)
vic_0000 <- sum((female_SCM$victim1 == 0) & (female_SCM$victim2 == 0) & 
                  (female_SCM$victim3 == 0) & (female_SCM$victim4 == 0), 
                na.rm = T)

vic_1111_o <- sum((female_outliers$victim1 == 1) & (female_outliers$victim2 == 1) & 
                  (female_outliers$victim3 == 1) & (female_outliers$victim4 == 1), 
                na.rm = T)
vic_0000_o <- sum((female_outliers$victim1 == 0) & (female_outliers$victim2 == 0) & 
                  (female_outliers$victim3 == 0) & (female_outliers$victim4 == 0), 
                na.rm = T)
@

2.	basic descriptive stats for outcome (eg if binary, how many events)

<<>>=
table_suicidal <- rbind(table(female_SCM$suicide1, useNA = "always"),
                        table(female_SCM$suicide2, useNA = "always"),
                        table(female_SCM$suicide3, useNA = "always"),
                        table(female_SCM$suicide4, useNA = "always"),
                        table(female_SCM$suicide5, useNA = "always"))
rownames(table_suicidal) <- paste0("yr", 1:5)
table(female_SCM$Y, useNA = "always")

table(female_outliers$Y, useNA = "always")

@

3.	“Table 1” descriptive (univariate) sample characteristics

<<>>=
table(female_SCM$race, female_SCM$cohort, useNA = "always")[, c(1, 2)]
table(female_SCM$sxorient, female_SCM$cohort, useNA = "always")[, c(1, 2)]
table(female_SCM$consent, female_SCM$cohort, useNA = "always")[, c(1, 2)]
table(female_SCM$forced, female_SCM$cohort, useNA = "always")[, c(1, 2)]
table(female_SCM$forced_att, female_SCM$cohort, useNA = "always")[, c(1, 2)]
table(female_SCM$sex_bef14, female_SCM$cohort, useNA = "always")[, c(1, 2)]
@


\end{table}

<<>>=
table_relation <- rbind(table(female_SCM$relation1, useNA = "always"),
                        table(female_SCM$relation2, useNA = "always"),
                        table(female_SCM$relation3, useNA = "always"),
                        table(female_SCM$relation4, useNA = "always"),
                        table(female_SCM$relation5, useNA = "always"))
rownames(table_relation) <- paste0("yr", 1:5)

table_religion <- rbind(table(female_SCM$religion1, useNA = "always"),
                        table(female_SCM$religion2, useNA = "always"),
                        table(female_SCM$religion3, useNA = "always"),
                        table(female_SCM$religion4, useNA = "always"),
                        table(female_SCM$religion5, useNA = "always"))
rownames(table_religion) <- paste0("yr", 1:5)

table_victfrnd <- rbind(table(female_SCM$victfrnd1, useNA = "always"),
                        table(female_SCM$victfrnd2, useNA = "always"),
                        table(female_SCM$victfrnd3, useNA = "always"),
                        table(female_SCM$victfrnd4, useNA = "always"),
                        table(female_SCM$victfrnd5, useNA = "always"))
rownames(table_victfrnd) <- paste0("yr", 1:5)

table_drug <- rbind(table(female_SCM$drug1, useNA = "always"),
                    table(female_SCM$drug2, useNA = "always"),
                    table(female_SCM$drug3, useNA = "always"),
                    table(female_SCM$drug4, useNA = "always"),
                    table(female_SCM$drug5, useNA = "always"))
rownames(table_drug) <- paste0("yr", 1:5)

table_alcohol <- rbind(table(female_SCM$alcohol1, useNA = "always"),
                       table(female_SCM$alcohol2, useNA = "always"),
                       table(female_SCM$alcohol3, useNA = "always"),
                       table(female_SCM$alcohol4, useNA = "always"),
                       table(female_SCM$alcohol5, useNA = "always"))
rownames(table_alcohol) <- paste0("yr", 1:5)

table_suicide <- rbind(table(female_SCM$suicide1, useNA = "always"),
                       table(female_SCM$suicide2, useNA = "always"),
                       table(female_SCM$suicide3, useNA = "always"),
                       table(female_SCM$suicide4, useNA = "always"),
                       table(female_SCM$suicide5, useNA = "always"))
rownames(table_suicide) <- paste0("yr", 1:5)

table_victim <- rbind(table(female_SCM$victim1, useNA = "always"),
                       table(female_SCM$victim2, useNA = "always"),
                       table(female_SCM$victim3, useNA = "always"),
                       table(female_SCM$victim4, useNA = "always"))
rownames(table_victim) <- paste0("yr", 1:4)
@


<<>>=
vic_1111 <- sum((female_SCM$victim1 == 1) & (female_SCM$victim2 == 1) & 
                  (female_SCM$victim3 == 1) & (female_SCM$victim4 == 1), na.rm = T)
vic_111 <- sum((female_SCM$victim1 == 1) & (female_SCM$victim2 == 1) & 
                 (female_SCM$victim3 == 1), na.rm = T)
vic_11 <- sum((female_SCM$victim1 == 1) & (female_SCM$victim2 == 1), na.rm = T)
vic_1 <- sum(female_SCM$victim1 == 1, na.rm = T)
vic_1111 / nrow(female_SCM)
vic_1111/vic_111
vic_111/vic_11
vic_11/vic_1

vic_0000 <- sum((female_SCM$victim1 == 0) & (female_SCM$victim2 == 0) & 
                  (female_SCM$victim3 == 0) & (female_SCM$victim4 == 0), na.rm = T)
vic_000 <- sum((female_SCM$victim1 == 0) & (female_SCM$victim2 == 0) & 
                  (female_SCM$victim3 == 0), na.rm = T)
vic_00 <- sum((female_SCM$victim1 == 0) & (female_SCM$victim2 == 0), na.rm = T)
vic_0 <- sum(female_SCM$victim1 == 0, na.rm = T)
vic_0000 / nrow(female_SCM)
vic_0000/vic_000
vic_000/vic_00
vic_00/vic_0
@



\end{document}
